+++ src/services/rhai/telegram.rs
@@ -5,6 +5,7 @@ use rhai::{Dynamic, Engine, Map, RegisterResultFn};
 
 use crate::prelude::*;
 use crate::services::rhai::FnResult;
+use crate::services::telegram::method_call::MethodCall;
 use crate::services::telegram::*;
 
 pub fn register_functions(engine: &mut Engine) {
@@ -12,17 +13,14 @@ pub fn register_functions(engine: &mut Engine) {
         "send_message",
         |this: &mut Telegram, unique_id: i64, text: &str, options: Map| -> FnResult {
             let mut options = options;
-            this.call::<TelegramMessage>(
-                &TelegramMethodCall::SendMessage {
+            into_fn_result(this.call::<TelegramMessage>(
+                &MethodCall::SendMessage {
                     chat_id: TelegramChatId::UniqueId(unique_id),
                     text: text.into(),
                     parse_mode: get_option(&mut options, "parse_mode"),
                 },
                 None,
-            )
-            .map(Dynamic::from)
-            .map_err(|error| error.to_string())
-            .map_err(Into::into)
+            ))
         },
     );
 
@@ -30,20 +28,31 @@ pub fn register_functions(engine: &mut Engine) {
         "send_video",
         |this: &mut Telegram, unique_id: i64, bytes: Arc<Bytes>, options: Map| -> FnResult {
             let mut options = options;
-            this.call::<TelegramMessage>(
-                &TelegramMethodCall::SendVideo {
-                    chat_id: TelegramChatId::UniqueId(unique_id),
-                    video: None, // TODO: override for `&str`.
-                    caption: get_option(&mut options, "caption"),
-                    parse_mode: get_option(&mut options, "parse_mode"),
-                },
+            into_fn_result(this.call::<TelegramMessage>(
+                &MethodCall::new_send_video(TelegramChatId::UniqueId(unique_id), None, &mut options),
                 Some(("video".into(), bytes)),
-            )
-            .map(Dynamic::from)
-            .map_err(|error| error.to_string())
-            .map_err(Into::into)
+            ))
         },
     );
+}
+
+impl MethodCall {
+    fn new_send_video(chat_id: TelegramChatId, video: Option<String>, options: &mut Map) -> Self {
+        MethodCall::SendVideo {
+            chat_id,
+            video,
+            caption: get_option(options, "caption"),
+            parse_mode: get_option(options, "parse_mode"),
+        }
+    }
+}
+
+/// Converts an application `Result<T>` into Rhai function result.
+fn into_fn_result<T: Sync + Send + Clone + 'static>(result: Result<T>) -> FnResult {
+    result
+        .map(Dynamic::from)
+        .map_err(|error| error.to_string())
+        .map_err(Into::into)
 }
 
 /// Extract the option from the map and safely cast to the specified type.